ci : add github release job
This commit adds a GitHub Actions workflow to automate the release
process. Currently this will only create an archive of the sources for
ggml when a tag is pushed.

The motivation for this is that when we start releasing versions of ggml
using semantic versioning it can be nice to have the sources needed for
ggml to be deployed as a github release. This enables CMake users that
use `FetchContent` efficiently specify the the zip file instead of
cloning:
```cmake
FetchContent_Declare(ggml URL https://ggml-org/ggml/download/ggml-src.zip)
FetchContent_MakeAvailable(ggml)
```